implement resource exception handling in TSResourceSource
Resource exceptions were not actually being handled.  This reworks how data
is pulled from the in queue to check if there are exceptions in the resource
(indicated by the reworked self.exception_signal), and then exit the
pipeline if any exceptions are encountered.

tests are added to maintain test coverage over the modified parts of the code.